requirements.txt
fastapi
uvicorn
streamlit
pycryptodome
requests
python-dotenv
playwright
langchain
scikit-learn 